// src/generated/models/Tk_customersModel.ts
/*!
 * Copyright (C) Microsoft Corporation. All rights reserved.
 * This file is autogenerated. Do not edit this file directly.
 */
export const Tk_customersstatecode = {
  0: 'Active',
  1: 'Inactive'
} as const;
export type Tk_customersstatecode = keyof typeof Tk_customersstatecode;
export const Tk_customersstatuscode = {
  1: 'Active',
  2: 'Inactive'
} as const;
export type Tk_customersstatuscode = keyof typeof Tk_customersstatuscode;
export const Tk_customerstk_title = {
  126790000: 'Mr_',
  126790001: 'Miss_',
  126790002: 'Ms_',
  126790003: 'Dr_'
} as const;
export type Tk_customerstk_title = keyof typeof Tk_customerstk_title;

export interface Tk_customersBase {
  importsequencenumber?: number;
  overriddencreatedon?: string;
  statecode: Tk_customersstatecode;
  statuscode?: Tk_customersstatuscode;
  timezoneruleversionnumber?: number;
  tk_customeraddress?: string;
  tk_customeremail?: string;
  tk_customerid: string;
  tk_customerref?: string;
  tk_customertelephone?: string;
  tk_firstname: string;
  tk_fullname: string;
  tk_lastname: string;
  tk_title?: Tk_customerstk_title;
  utcconversiontimezonecode?: number;
}

export interface Tk_customers extends Tk_customersBase {
  createdbyname?: string;
  createdbyyominame: string;
  createdon?: string;
  createdonbehalfbyname?: string;
  createdonbehalfbyyominame: string;
  modifiedbyname?: string;
  modifiedbyyominame: string;
  modifiedon?: string;
  modifiedonbehalfbyname?: string;
  modifiedonbehalfbyyominame: string;
  organizationidname: string;
  statecodename?: string;
  statuscodename?: string;
  tk_titlename?: string;
  versionnumber?: number;
  createdby?: object;
  _createdby_value?: string;
  createdonbehalfby?: object;
  _createdonbehalfby_value?: string;
  modifiedby?: object;
  _modifiedby_value?: string;
  modifiedonbehalfby?: object;
  _modifiedonbehalfby_value?: string;
  organizationid?: object;
  _organizationid_value?: string;
}